+++ src/keret-controller/src/domain/port.rs
@@ -1,18 +1,8 @@
 use crate::{
-    domain::{Duration, Instant},
+    domain::model::{Duration, Instant, InteractionRequest},
     error::Error,
 };
 use tiny_led_matrix::Render;
-
-/// enum to indicate the users desired interaction
-/// which is calculated by which button was pressed
-#[derive(Debug, Copy, Clone, Default)]
-pub(crate) enum InteractionRequest {
-    #[default]
-    None,
-    ToggleMode,
-    Reset,
-}
 
 pub(crate) trait RunningTimeClock {
     fn now(&mut self) -> Instant;
